--- src/main/java/com/spring/admin/space/repository/SpaceDetailRepository.java
@@ -1,5 +1,0 @@
-package com.spring.admin.space.repository;
-
-public class SpaceDetailRepository {
-
-}
