+++ src/main/resources/static/css/admin/adminMain.css
@@ -3,42 +3,6 @@
     height: 100vh;
 }
 
-.sidebar {
-    width: 250px;
-    background-color: #343a40;
-    color: white;
-    padding: 15px;
-    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
-}
-
-.sidebar h2 {
-    text-align: center;
-    margin-bottom: 20px;
-    font-size: 1.5rem;
-}
-
-.sidebar ul {
-    list-style-type: none;
-    padding: 0;
-}
-
-.sidebar ul li {
-    margin: 15px 0;
-}
-
-.sidebar ul li a {
-    color: white;
-    text-decoration: none;
-    display: block;
-    padding: 10px;
-    border-radius: 5px;
-    transition: background-color 0.3s;
-}
-
-.sidebar ul li a:hover {
-    background-color: #495057;
-}
-
 /* 메인 콘텐츠 스타일 */
 .container {
     padding: 20px;
