+++ src/main/resources/static/css/admin/space/adminSpaceList.css
@@ -1,0 +1,1 @@
+@charset "UTF-8";
