--- src/main/resources/static/css/adminMain.css
@@ -1,93 +1,0 @@
-.wrapper {
-    display: flex;
-    height: 100vh;
-}
-
-.sidebar {
-    width: 250px;
-    background-color: #343a40;
-    color: white;
-    padding: 15px;
-    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
-}
-
-.sidebar h2 {
-    text-align: center;
-    margin-bottom: 20px;
-    font-size: 1.5rem;
-}
-
-.sidebar ul {
-    list-style-type: none;
-    padding: 0;
-}
-
-.sidebar ul li {
-    margin: 15px 0;
-}
-
-.sidebar ul li a {
-    color: white;
-    text-decoration: none;
-    display: block;
-    padding: 10px;
-    border-radius: 5px;
-    transition: background-color 0.3s;
-}
-
-.sidebar ul li a:hover {
-    background-color: #495057;
-}
-
-/* 메인 콘텐츠 스타일 */
-.container {
-    padding: 20px;
-    max-width: 1200px; /* 사용자 페이지와 동일한 너비 설정 */
-    margin: 0 auto; /* 중앙 정렬 */
-}
-
-h2 {
-    text-align: center; /* 제목 중앙 정렬 */
-    margin-top: 0;
-}
-
-.dashboard {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    gap: 20px;
-    margin-top: 20px;
-}
-
-.card {
-    background-color: white;
-    padding: 20px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    border-radius: 8px;
-    flex: 1;
-    min-width: 300px; /* 카드 최소 너비 설정 */
-    text-align: center; /* 카드 내용 중앙 정렬 */
-}
-
-.card h3 {
-    margin: 0 0 10px;
-    font-size: 18px;
-}
-
-.card p {
-    margin: 0;
-    font-size: 16px;
-}
-
-.inquiry-list, .popular-spaces {
-    list-style-type: none;
-    padding: 0;
-    text-align: left; /* 리스트는 왼쪽 정렬 유지 */
-}
-
-.inquiry-list li, .popular-spaces li {
-    margin-bottom: 10px;
-    padding: 5px;
-    background-color: #f9f9f9;
-    border-radius: 4px;
-}
