+++ src/main/resources/templates/template/admin/fragments/adminNav.html
@@ -1,0 +1,48 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+
+	<th:block th:fragment="config">
+		<script src="/assets/js/color-modes.js"></script>
+		<meta charset="UTF-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+		<title> SpaceHub - 관리자페이지</title>
+			
+		<link rel="shortcut icon" href="/images/icon.png" />
+		<link rel="apple-touch-icon" href="/images/icon.png" />
+			
+		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@docsearch/css@3">
+		
+		<!-- <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css" > -->
+		<link href="/assets/dist/css/bootstrap.min.css" rel="stylesheet">
+		<link type="text/css" rel="stylesheet" href="/css/client/common.css" />
+		<!-- Custom styles for this template -->
+    	<link href="/css/admin/adminNav.css" rel="stylesheet">
+    	
+	</th:block>
+	
+
+    <!-- 관리자 페이지 왼쪽 네비게이션 -->
+    <th:block th:fragment="admin-nav">
+    	<body>
+			<div class="nav-container">
+			    <nav class="sidebar">
+			        <h2>SpaceHub <br> 관리자페이지</h2>
+			        <ul>
+			            <li><a href="/admin/main">홈</a></li>
+			            <li><a href="/admin/myPage">관리자 마이페이지</a></li>
+			            <li><a href="/admin/space/spaceList">공간 관리</a></li>
+			            <li><a href="#">리뷰 관리</a></li>
+			            <li><a href="#">문의 게시판</a></li>
+			            <li><a href="#">공지 게시판</a></li>
+			            <li><a href="/admin/memberManage">회원 관리</a></li>
+			        </ul>
+			    </nav>
+			</div>
+        </body>
+    </th:block>
+
+
+
+</html>
+
